prevent unlink for other district

+++ care/users/api/viewsets/users.py
@@ -286,6 +286,9 @@ class UserViewSet(
             )
         if user.home_facility == facility:
             raise ValidationError({"facility": "Cannot Delete User's Home Facility"})
+        # ensure that district admin only able to delete in the same district
+        if user.district.name != requesting_user.district.name:
+            raise ValidationError({"facility": "Cannot delete User's Facility from other district"})
         FacilityUser.objects.filter(facility=facility, user=user).delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
